feat(PackageDelivery (MultiWays)): New feature for Driver Android - Multi Ways payment method enhancement.
refers to LB-16168

+++ app/src/main/java/com/logestechs/driver/api/responses/GetDeficitBalanceResponse.kt
@@ -6,5 +6,5 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class GetDeficitBalanceResponse(
     var id: Long,
-    var amount: Int? = null,
+    var amount: Double? = null,
 ) : Parcelable
